requirements.txt
imageio==2.36.0
numpy==2.1.3
pillow==11.0.0
